+++ oscmd.go
@@ -8,10 +8,14 @@ import (
 
 // NewOSCmd returns a command from the OS values.
 func NewOSCmd() *OSCmd {
-	return &OSCmd{}
+	return &OSCmd{
+		exit: os.Exit,
+	}
 }
 
-type OSCmd struct{}
+type OSCmd struct {
+	exit func(code int)
+}
 
 func (me *OSCmd) Getenv(key string) string { return os.Getenv(key) }
 func (me *OSCmd) Args() []string           { return os.Args }
@@ -20,6 +24,9 @@ func (me *OSCmd) Stdin() io.Reader         { return os.Stdin }
 func (me *OSCmd) Stdout() io.Writer        { return os.Stdout }
 func (me *OSCmd) Stderr() io.Writer        { return os.Stderr }
 
-// Stop returns the given exit code
-func (me *OSCmd) Stop(exitCode int) int { return exitCode }
-func (me *TCmd) Fatal(v ...interface{}) { log.Fatal(v...) }
+// Exit returns the given exit code
+func (me *OSCmd) Exit(code int) { me.exit(code) }
+func (me *OSCmd) Fatal(v ...interface{}) {
+	log.Println(v...)
+	me.exit(1)
+}