+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Renamed Command.Stop to Exit
 - Add Command.Fatal
 
 ## [0.6.1] 2021-03-21